3D Random Sinusous HTML
This will be a little hard to explain, so please bare with me. This is a 3D graph on a 2D canvas, this works based on each area's HEX value. So in each area that is black that is the lowest point, and every place that is white that is the highest point, think of them as -1 and +1. I'm currently working on rewriting this since I can compress it's content more.